Add green/red diff highlighting in git commit msg
When viewing a diff via `git commit -v`, the highlighting that is passed
to Vim is that of the color scheme.  What this does is it modifies it to
look more like the default diff colors from `git diff` (using green for
additions and red for subtractions).  This makes viewing the diff more
obvious.

That said, this, and many or the other color additions here should be
done elsewhere, but that is for another time.