Fix ruby snippets to use new DIR structure
The new maintainer of vim-snipmate changed how subfiles for snippets
work, and now use a subdirectory for language specific files.  This
simply renames those files to work with the new structure.

Also added the `def` snippet for ruby, since it didn't exist for some
reason (and I use it ALL the time).